funkcna obrazovka zobrazenia otazok konkretneho tagu - zatial nedokonceny actionBar

+++ app/src/main/java/com/example/mtaafe/data/models/QuestionsList.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class QuestionsList (
         @SerializedName("count")
-        var count: Int,
+        var count: Long,
 
         @SerializedName("questions")
         var questions: ArrayList<QuestionItem>
